Create HomeUtils and PatientUtils, and refactor PatientPrimarySpec
