+++ src/main/java/duke/command/CommandUtils.java
@@ -1,13 +1,5 @@
 package duke.command;
 
-import duke.DukeCore;
-import duke.data.DukeObject;
-import duke.data.Patient;
-import duke.data.SearchResults;
-import duke.exception.DukeException;
-import duke.exception.DukeUtilException;
-
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -209,102 +201,6 @@ public class CommandUtils {
     }
 
     // TODO might move these to respective DukeObjects
-
-    /**
-     * Find a {@code Patient} with the supplied identifier. Only 1 of either bed number or displayed index
-     * should be used to identify said patient.
-     *
-     * @param core  DukeCore object.
-     * @param bedNo Bed number of patient.
-     * @param nameOrIdx Displayed index of patient (Home context), or name of Patient.
-     * @return Patient object
-     * @throws DukeException If 1 of the following 3 conditions applies.
-     *                       1. No identifier is provided.
-     *                       2. 2 identifiers are provided.
-     *                       3. 1 unique identifier is provided but said patient does not exist.
-     */
-    public static Patient findFromHome(DukeCore core, String bedNo, String nameOrIdx) throws DukeException {
-        if (nameOrIdx == null && bedNo == null) {
-            throw new DukeUtilException("Please provide a way to identify the patient! Patients can be searched for"
-                    + "by name/index or by bed.");
-        }
-        if (nameOrIdx != null && bedNo != null) {
-            throw new DukeUtilException("I don't know if you want me to find the patient ");
-        }
-        if (bedNo != null) {
-            return core.patientData.getPatientByBed(bedNo);
-        }
-
-        int index = idxFromString(nameOrIdx);
-        if (index != -1) {
-            // TODO: Law of demeter
-            List<DukeObject> patientList = core.ui.getIndexedList("patient");
-            int count = patientList.size();
-            if (index > count) {
-                throw new DukeException("I have only " + ((count == 1) ? ("1 patient") : (count + "patients")) + " in "
-                        + "my list!");
-            }
-            return (Patient) patientList.get(index - 1);
-        } else {
-            return null;
-        }
-    }
-
-    // TODO clarify if we are doing get by idx from the UI or from the data
-
-    /**
-     * Find a {@code DukeObject} with the supplied identifier. Only 1 of either name or displayed index should be used
-     * to identify said DukeObject.
-     *
-     * @param core    DukeCore object.
-     * @param patient Patient object.
-     * @param type    Type of DukeObject.
-     * @param nameOrIdx   Name of DukeObject or displayed index.
-     * @return DukeObject object,
-     * @throws DukeException If 1 of the following 3 conditions applies.
-     *                       1. No identifier is provided.
-     *                       2. 2 identifiers are provided.
-     *                       3. 1 unique identifier is provided but said DukeObject does not exist.
-     */
-    public static DukeObject findFromPatient(DukeCore core, Patient patient, String type, String nameOrIdx)
-            throws DukeException {
-        int index = idxFromString(nameOrIdx);
-        if (index != -1) {
-            if (type == null) {
-                throw new DukeUtilException("I don't know which list you want me to get from!");
-            }
-            try {
-                return core.ui.getIndexedList(type).get(index - 1);
-            } catch (IndexOutOfBoundsException e) {
-                throw new DukeException("I don't have a " + type + " of that index!");
-            }
-        } else {
-            return null;
-        }
-    }
-
-
-    public static SearchResults searchFromPatient(Patient patient, String type, String searchTerm) throws DukeUtilException {
-        SearchResults results = new SearchResults(searchTerm, new ArrayList<DukeObject>(), patient);
-        if (type == null) {
-            results.addAll(patient.findImpressionsByName(searchTerm));
-            results.addAll(patient.findCriticalsByName(searchTerm));
-            results.addAll(patient.findFollowUpsByName(searchTerm));
-        }
-
-        if ("impression".equals(type)) {
-            results.addAll(patient.findImpressionsByName(searchTerm));
-        } else if ("critical".equals(type)) {
-            results.addAll(patient.findCriticalsByName(searchTerm));
-        } else {
-            results.addAll(patient.findFollowUpsByName(searchTerm));
-        }
-
-        if (results.getCount() == 0) {
-            throw new DukeUtilException("Can't find anything with those search parameters!");
-        }
-        return results;
-    }
 
 
     /**
