+++ src/main/java/duke/command/home/HomeObjSpec.java
@@ -6,13 +6,16 @@ import duke.command.ObjSpec;
 import duke.data.Patient;
 import duke.exception.DukeException;
 
+/**
+ * Template class for Home context commands that require a patient to be found by bed, index or name.
+ */
 public abstract class HomeObjSpec extends ObjSpec {
 
     @Override
     protected void execute(DukeCore core) throws DukeException {
         super.execute(core);
         String bed = cmd.getSwitchVal("bed");
-        Patient patient = CommandUtils.findFromHome(core, bed, cmd.getArg());
+        Patient patient = HomeUtils.findFromHome(core, bed, cmd.getArg());
         if (patient == null) {
             core.search(core.patientData.findPatients(cmd.getArg()), cmd);
         } else {
