+++ src/main/java/duke/command/impression/ImpressionObjSpec.java
@@ -1,0 +1,4 @@
+package duke.command.impression;
+
+public class ImpressionObjSpec {
+}
